Wait for build to start before streaming the console
Ensure that when you trigger a job like:

    nestor build --console myjob

the console output is associated with the newly created build (and not
the previous build) even if the newly created build takes longer than 5
seconds to start.

This change involves using the 'Location' returned when a job is
triggered to monitor the new build and proceed to stream the console
only when the API indicates that the build has started. nestor will poll
the monitoring URL every 2 seconds, up to a maximum of 30 seconds (when
it will give up).